Fix up cloning for SolidFire volumes
There were a few things missing in the clone implementation for
SolidFire, most importantly we weren't setting up the attributes for
things like the ndvp versioning and the original docker-name that is
used to find volumes.

The SF Clone requires that you set attributes seperately, so we add that
here as a second step in the clone process.  Another feature to consider
is allowing to set new QoS values on a clone, we'll add that in a follow
up patch.