Add counter to mounts for SolidFire
SolidFire is a bit different than some other devices in the portfolio
in that it uses a unique target for each volume and it doesn't do
any internal reference counting in terms of attachment counts.

This gets pretty wonky in Docker becuase docker does a lot of
mount/unmount calls in the background, particularly when using
commands like `docker cp ./some-data container-foo:/data/`

It's debatable whether docker is well behaved here, but regardless
we can and should do better in terms of handling things like this.

This change adds a simple mechanism where we count the number of
Mounts issued against a volume.  We increment this count each time
a new mount request is received, and decrement it each time an
unmount request is received.  This way we don't accidentally rip
a volume out from under a container during things like docker cp
and more importantly we don't blow up if somebody is trying to do
multiple attachments to a single volume from multiple containers on
the same node.

For now this isn't persistent, and it can also cause some difficulty
with hanging connections if the counter somehow gets out of sync.  Those
types of things should be considered/addressed in the future, and we
will probably need somebody to get on creating an admin cli like the
solidfire-docker-driver has to do clean ups etc.

Addresses Github Issue: #93