Refrain from unmounting until volume removal
There are some questionable behaviors on the docker side wrt
sequences of mounting and unmounting.  The `docker cp` command
in particular has a good bit of seemingly unnecessary mount/unmount
operations that for some devices (those that use unique targets)
can trip things up a bit.

The problems include removing a device that's in use by a container
as well as removing an attachment that is expected to be present.

One way of dealing with this (and it's called out in the Docker
Volume API docs) is to utilize reference counting of the mounts
and track things that way.  We could do that, but there's some
concern around that in terms of things getting out of sync and
also storing state in the driver isn't the most ideal thing either.

It turns out however that maybe none of this is really necessary.
Instead of always unmounting and removing iSCSI connections we can
instead leave mounts and iscsi connections in place for the life
of a volume.

So what this change does is makes the Docker API Unmount basically
a NOOP.  We still do things like Mount and create iSCSI connection
the same as before, but now we ignore the Unmount command.  If
subsequent or multiple Mount requests come in for a volume that's
fine, we just reuse the connection and mount point that we already
have in place.  We leave this in place until the volume is removed.

This works will in the case of things like reboots, or restarts of
NDVP as well, because we don't need to worry about state or ref
counting.  If a Mount cmd is received in those scenarios and the
mount was lost due to a reboot, we just create one and go about
our business.

Addresses Github Issue: #93