Add SolidFire Redis Queue Create Opt
One problem with a decentralized scale-out architecture
is it's possible to receive duplicate commands from the
various clients.

Swarm is a good example when using UDP, a Create call is
issued by UDP to ALL nodes in the SWARM, SolidFire doesn't
have any rules against duplicate names, and it also has a
bit of a race internally between receiving the Create cmd
and actually returning the volume on a Get call.

The result is that in a large Swarm cluster you end up
creating several duplicate volumes.

There are a number of ways to design around this, however the
simplest solution is to insert a Queue/Controller layer to
sort of direct/route these requests.

This patch adds an option to configure a Redis server/container
to act as a controller for Creates.  If configured, we'll simply
attempt to set a K/V in the Redis DB upon Create if we don't
discover the Volume on the cluster already, and then continue
on and issue the create.

If the Redis command returns that an entry exists already (ie
create is in process), then we'll just wait a couple (?) seconds
attempt a Get call to the cluster again and hopefully the volume
will be reported by SolidFire by then, and we'll return that the
volume already exists.

We also set an expiration on the Redis entry, so it doesn't hang
around forever.  The Redis container dies?  No problem, just spawn
another one.  In fact, you *SHOULD* make the Redis Container an
item in your Swarm cluster and let Swarm do all the hard work.

We should run some tests on these times, and perhaps consider
making them configurable.